Rename noteburstjob domain module to noteburst
This will allow us to add a client to this module and generalize it from
only being models from noteburst responses created elsewhere.
