+++ src/timessquare/domain/noteburst.py
